Additional changes to fix circular dependencies and linting
- exported interfaces and injection tokens which caused circular dependency errors
- bumped angular-eslint packages to 13.5.0
