+++ package.json
@@ -52,11 +52,11 @@
     "@actions/github": "2.2.0",
     "@angular-devkit/architect": "^0.1303.8",
     "@angular-devkit/build-angular": "13.3.8",
-    "@angular-eslint/builder": "1.2.0",
-    "@angular-eslint/eslint-plugin": "1.2.0",
-    "@angular-eslint/eslint-plugin-template": "1.2.0",
-    "@angular-eslint/schematics": "1.2.0",
-    "@angular-eslint/template-parser": "1.2.0",
+    "@angular-eslint/builder": "^13.5.0",
+    "@angular-eslint/eslint-plugin": "^13.5.0",
+    "@angular-eslint/eslint-plugin-template": "^13.5.0",
+    "@angular-eslint/schematics": "^13.5.0",
+    "@angular-eslint/template-parser": "^13.5.0",
     "@angular/cli": "13.3.8",
     "@angular/compiler-cli": "13.3.11",
     "@angular/language-service": "13.3.11",
